update to v1.9-pre.14

+++ Entity/Pathfinding/StraightLine.cs
@@ -15,8 +15,8 @@ namespace Vintagestory.API.Common
         {
             if (entity?.Properties.Server?.Attributes?.GetTreeAttribute("pathfinder") != null)
             {
-                minTurnAnglePerSec = (float)entity.Properties.Server.Attributes.GetTreeAttribute("pathfinder").GetFloat("minTurnAnglePerSec", 250);
-                maxTurnAnglePerSec = (float)entity.Properties.Server.Attributes.GetTreeAttribute("pathfinder").GetFloat("maxTurnAnglePerSec", 450);
+                minTurnAnglePerSec = (float)entity.Properties.Server.Attributes.GetTreeAttribute("pathfinder").GetDecimal("minTurnAnglePerSec", 250);
+                maxTurnAnglePerSec = (float)entity.Properties.Server.Attributes.GetTreeAttribute("pathfinder").GetDecimal("maxTurnAnglePerSec", 450);
             } else
             {
                 minTurnAnglePerSec = 250;
